requirements.txt
fastapi~=0.103.1
uvicorn==0.23.2
starlette~=0.27.0
python-dotenv~=1.0.0
PyYAML~=6.0.1
requests~=2.31.0
mlflow~=2.7.0
python-multipart==0.0.6
pandas~=2.1.3
pydantic~=1.10.13
boto3~=1.33.11
redis~=5.0.1
torch
pickle
